key pressing and dash implemented
I updated DirectionInput.js to work with the keypress event for keys like space, enter, or plus. This is stored in the keysPressed array, which is cleared every frame by the step function. For now, I pass the boolean of whether or not the space button is pressed as an argument for player.update, and from this, we can determine if he's dashing or not. If he is, we boost his speed and movingProgressRemaining, and I made some tweeks to the player movement functions that are not final because dashing doesn't work while he's moving.