level cells created
Implemented the basic setup of a level array in ship.js, which temporarily reads through and draws it onscreen for reference. This would be handy for player and enemy collisions and later when we do a pathfinding algorithm, because many things will be object based, but it would be tedious to declare wall or hole or lava or whatever objects because there's just so much of them in each scene. objects like chests, potions, money, enemies, or otherwise would be easier to handle object-based, and their placement would also be easier with this level array.