Dash Collider
Dash collider finally works. it operates by checking how many spaces are available in the dash() function and setting the dashingDistance equal to that number of spaces. A couple of hacky tweeks had to be added to the updatePosition function to make sure that the player still stayed centered in a cell if the dashingDistance was not divisible by three (this.speed). included this.map of game.js with an init function to all game objects so that it doesn't have to be passed to them in the update function and can be used in their code for other functions (such as dash()).