dash wall creation glitch fix
there was a glitch where a wall would be left behind when the player dashed because he started moving normally and never deleted the wall in the move function when he dashed. this has been fixed (in the dash() function) as it removes a wall at the beginnning of the dash relative to the player's moving progress remaining.